# go-dbu-api-develop/docs/swagger.yaml
basePath: /
definitions:
...
  models.AssignedGood:
    properties:
      code:
        type: string
    type: object
  models.AssignmentRoom:
    properties:
      estado:
        type: string
      fecha_asignacion:
        type: string
      id:
        type: string
    type: object
...
  models.Links:
    properties:
      next:
        type: string
      prev:
        type: string
    type: object
...
  models.Metadata:
    properties:
      limit:
        type: integer
      links:
        $ref: '#/definitions/models.Links'
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      metadata:
        $ref: '#/definitions/models.Metadata'
      msg:
        type: string
      type:
        type: string
    type: object
  models.Room:
    properties:
      capacity:
        type: integer
      id:
        type: string
      number:
        type: integer
      status:
        type: string
    type: object
  models.RoomMate:
    properties:
      code:
        type: string
      full_name:
        type: string
    type: object
...
  models.Sanction:
    properties:
      date:
        type: string
      description:
        type: string
    type: object
...
  models.Student:
    properties:
      assigned_goods:
        items:
          $ref: '#/definitions/models.AssignedGood'
        type: array
      room_mates:
        items:
          $ref: '#/definitions/models.RoomMate'
        type: array
      sanctions:
        items:
          $ref: '#/definitions/models.Sanction'
        type: array
      student:
        $ref: '#/definitions/models.StudentInfo'
    type: object
  models.StudentInfo:
    properties:
      admission_date:
        type: string
      code:
        type: string
      department:
        type: string
      faculty:
        type: string
      full_name:
        type: string
      id:
        type: integer
      number_identification:
        type: string
      professional_school:
        type: string
      residence:
        type: string
      room:
        $ref: '#/definitions/models.Room'
      sex:
        type: string
    type: object
...
  rooms.AssignmentRoomRequest:
    properties:
      student_id:
        type: integer
      submission_id:
        type: integer
    type: object
  rooms.DeleteAssignmentRoomRequest:
    properties:
      observation:
        type: string
      status:
        type: string
      student_id:
        type: integer
      submission_id:
        type: integer
    type: object
  rooms.RoomRequest:
    properties:
      capacity:
        type: integer
      id:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: juanm.campos@unas.edu.pe
    name: API Support
  description: DBU API
  license:
    name: Comercial
    url: http://bjungle.net/licenses
  termsOfService: https://dbu.com/terms/
  title: DBU API
  version: "1.0"
paths:
...
  /v1/residencias/{id}/cuartos/:
    get:
      consumes:
      - application/json
      description: Método que permite obtener todos los alumnos asociados a una Residencia
      parameters:
      - description: ID de la Residencia
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Número de página
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Límite de registros por página
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: ID de la solicitud
        in: query
        name: submission_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Alumnos obtenidos exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data:
                  items:
                    $ref: '#/definitions/models.Student'
                  type: array
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "400":
          description: Error en la solicitud
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "401":
          description: Error de autenticación
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Obtener todos los alumnos de una Residencia
      tags:
      - Residencias
  /v1/residencias/cuartos:
    put:
      consumes:
      - application/json
      description: Método que permite Actualiza una instancia del objeto cuartos en
        la base de datos
      parameters:
      - description: Datos para actualizar cuartos
        in: body
        name: RoomRequest
        required: true
        schema:
          $ref: '#/definitions/rooms.RoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Actualiza una instancia de cuartos
      tags:
      - Cuartos
  /v1/residencias/cuartos/{id}/asignar:
    post:
      consumes:
      - application/json
      description: Asigna un estudiante a un cuarto específico durante una convocatoria
        activa
      parameters:
      - description: ID del cuarto
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos de asignación del estudiante
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rooms.AssignmentRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Asignación exitosa
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/models.AssignmentRoom'
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "201":
          description: Asignación creada exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/models.AssignmentRoom'
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "400":
          description: 'Error en la solicitud: puede incluir errores de validación,
            habitación sin capacidad (código 93), estudiante ya asignado (código 94),
            estudiante con otra asignación (código 95), o estudiante no aceptado (código
            96)'
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "401":
          description: Error de autenticación
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "403":
          description: Permisos insuficientes
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "404":
          description: Recurso no encontrado
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Asignar estudiante a un cuarto
      tags:
      - Cuartos
  /v1/residencias/cuartos/{id}/eliminar-asignacion:
    delete:
      consumes:
      - application/json
      description: Elimina la asignación de un estudiante a un cuarto específico
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del cuarto
        in: path
        name: id
        required: true
        type: string
      - description: Datos de la asignación a eliminar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rooms.DeleteAssignmentRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Eliminación exitosa
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "400":
          description: Error en la solicitud
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "401":
          description: Error de autenticación
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "403":
          description: Permisos insuficientes
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "404":
          description: Recurso no encontrado
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Eliminar asignación de cuarto
      tags:
      - Cuartos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"